requirements.txt
asgiref==3.5.0
certifi==2022.5.18.1
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
cryptography==37.0.2
cycler==0.11.0
dateparser==1.1.1
distlib==0.3.4
Django==4.0.3
django-mathfilters==1.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
filelock==3.7.0
fonttools==4.33.3
idna==3.3
invoice2data==0.3.6
kiwisolver==1.4.3
matplotlib==3.5.2
mysqlclient==2.1.0
numpy==1.22.4
opencv-python==4.6.0.66
packaging==21.3
pdfminer==20191125
pdfminer.six==20220319
pdfplumber==0.6.2
Pillow==9.1.0
platformdirs==2.5.2
pycparser==2.21
pycryptodome==3.14.1
PyJWT==2.4.0
pyparsing==3.0.9
PyPDF2==1.27.12
PyPDF4==1.27.0
pytesseract==0.3.9
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
razorpay==1.3.0
regex==2022.3.2
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.2
Unidecode==1.3.4
urllib3==1.26.9
virtualenv==20.14.1
Wand==0.6.7
